+++ fili-core/src/main/java/com/yahoo/bard/webservice/data/config/table/PermissivePhysicalTableDefinition.java
@@ -9,25 +9,25 @@ import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.metadata.DataSourceMetadataService;
 import com.yahoo.bard.webservice.table.ConfigPhysicalTable;
-import com.yahoo.bard.webservice.table.PermissiveConcretePhysicalTable;
+import com.yahoo.bard.webservice.table.PermissivePhysicalTable;
 
 import java.util.Map;
 import java.util.Set;
 
 /**
- * Holds the fields needed to define a Permissive Concrete Physical Table.
+ * Holds the fields needed to define a Permissive Physical Table.
  */
-public class PermissiveConcretePhysicalTableDefinition extends ConcretePhysicalTableDefinition {
+public class PermissivePhysicalTableDefinition extends ConcretePhysicalTableDefinition {
 
     /**
-     * Define a permissive concrete physical table.
+     * Define a permissive physical table.
      *
      * @param name  The table name
      * @param timeGrain  The zoned time grain
      * @param metricNames  The Set of metric names on the table
      * @param dimensionConfigs  The dimension configurations
      */
-    public PermissiveConcretePhysicalTableDefinition(
+    public PermissivePhysicalTableDefinition(
             TableName name,
             ZonedTimeGrain timeGrain,
             Set<FieldName> metricNames,
@@ -37,7 +37,7 @@ public class PermissiveConcretePhysicalTableDefinition extends ConcretePhysicalT
     }
 
     /**
-     * Define a permissive concrete physical table with provided logical to physical column name mappings.
+     * Define a permissive physical table with provided logical to physical column name mappings.
      *
      * @param name  The table name
      * @param timeGrain  The zoned time grain
@@ -45,7 +45,7 @@ public class PermissiveConcretePhysicalTableDefinition extends ConcretePhysicalT
      * @param dimensionConfigs  The dimension configurations
      * @param logicalToPhysicalNames  A map from logical column names to physical column names
      */
-    public PermissiveConcretePhysicalTableDefinition(
+    public PermissivePhysicalTableDefinition(
             TableName name,
             ZonedTimeGrain timeGrain,
             Set<FieldName> metricNames,
@@ -62,7 +62,7 @@ public class PermissiveConcretePhysicalTableDefinition extends ConcretePhysicalT
             ResourceDictionaries dictionaries,
             DataSourceMetadataService metadataService
     ) {
-        return new PermissiveConcretePhysicalTable(
+        return new PermissivePhysicalTable(
                 getName(),
                 getTimeGrain(),
                 buildColumns(dictionaries.getDimensionDictionary()),